Optimizing error handling

+++ src/services/userService.js
@@ -1,5 +1,8 @@
 const userRepository = require('./../dataAccess/userRepository');
 const transformer = require('./../transformers/userTransformer');
+const { ForbiddenError } = require('./../exceptions/customErrors');
+
+let currentUser;
 
 async function authenticate({ userName, password }) {
     const user = await userRepository.getForAuth(userName, password);
@@ -13,6 +16,10 @@ async function getAll(showProtected) {
 }
 
 async function addUser(user) {
+    if (currentUser && !currentUser.isAdmin) {
+        throw new ForbiddenError('Function available to admins only');
+    }
+
     if (!user) return;
 
     return transformer.transform(await userRepository.addUser(user), true);
@@ -22,12 +29,24 @@ async function getUser(id, showProtected) {
     return transformer.transform(await userRepository.getUser(id), showProtected);
 }
 
-function removeUser(id) {
+async function removeUser(id) {
+    if (currentUser && !currentUser.isAdmin) {
+        throw new ForbiddenError('Function available to admins only');
+    }
+
     return userRepository.removeUser(id);
 }
 
-function updateUser(id, user) {
+async function updateUser(id, user) {
+    if (currentUser && !currentUser.isAdmin) {
+        throw new ForbiddenError('Function available to admins only');
+    }
+
     return userRepository.updateUser(id, user);
+}
+
+function setCurrentUser(user) {
+    currentUser = user;
 }
 
 module.exports = {
@@ -37,4 +56,5 @@ module.exports = {
     getUser,
     removeUser,
     updateUser,
+    setCurrentUser,
 };
